% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initParams.R
\name{initParams}
\alias{initParams}
\title{Initialize the parameters for the EM algorithm}
\usage{
initParams(data)
}
\arguments{
\item{data}{a matrix with the columns as attributes and rows as observations}
}
\value{
params - a list with mu and Sigma
}
\description{
Initialize the parameters for the EM algorithm
}
\examples{
nCol <- 10
nRow <- 100
data <- matrix(runif(nCol * nRow),nrow=nRow,ncol=nCol)
data[1,1:9] <- NA

params <- initParams(data)
}
