% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingNormalImpute.R
\name{missingNormalImpute}
\alias{missingNormalImpute}
\title{Impute the missing data using the parameters params}
\usage{
missingNormalImpute(params, data)
}
\arguments{
\item{params}{a list with mu and Sigma; it should be the output from the runEm function}

\item{data}{a matrix with the columns as attributes and rows as observations}
}
\value{
the data, but with the NA's imputed
}
\description{
Impute the missing data using the parameters params
}
\examples{
nCol <- 10
nRow <- 100
data <- matrix(runif(nCol * nRow),nrow=nRow,ncol=nCol)
data[1,1:9] <- NA

maxIterations <- 10
params <- runEm(data,maxIterations,howToStop = "params")
newData <- missingNormalImpute(params, data)
}
