% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEm.R
\name{runEm}
\alias{runEm}
\title{Run the EM algorithm}
\usage{
runEm(data, maxIterations, oldParams = NULL, howToStop = "loglike",
  eps = 1e-06)
}
\arguments{
\item{data}{a matrix with the columns as attributes and rows as observations}

\item{maxIterations}{the maximum number of iterations for the EM algorithm}

\item{oldParams}{the initialization of parameters: a list with mu and Sigma; if NULL, then we initialize the parameters via the function initParams}

\item{howToStop}{"loglike" or "params"; check whether the loglikelihood does not change from an iteration to another OR the parameters do not change}

\item{eps}{used to compare the loglikelihoods/params from two consecutive iterations; if the relative error is < eps then STOP}
}
\value{
params - a list with mu and Sigma - after running the EM algorithm
}
\description{
This function runs the Expectation Maximization algorithm for obtaining the MLE estimates for a multivariate normal distribution when there are missing values in data.
}
\examples{
nCol <- 10
nRow <- 100
data <- matrix(runif(nCol * nRow),nrow=nRow,ncol=nCol)
data[1,1:9] <- NA

params <- initParams(data)
maxIterations <- 10
params <- runEm(data,maxIterations,oldParams=params,howToStop = "loglike",eps = 1e-6)

# OR
nCol <- 10
nRow <- 100
data <- matrix(runif(nCol * nRow),nrow=nRow,ncol=nCol)
data[1,1:9] <- NA

maxIterations <- 10
params <- runEm(data,maxIterations,howToStop = "params")
}
